+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [0.0.2] - 2022-11-26
 
+### Changed
+- bubble_sort from function to class. It now calculates metadata and animation 
+sequences which we can get after the algorithm has finished running.
+
 ### Added
 - insertion_sort.py module and tests
 - selection_sort.py module and tests